Move classic implementation into separate package
