+++ src/main/java/ru/bulldog/justmap/map/data/classic/ChunkData.java
@@ -1,7 +1,9 @@
-package ru.bulldog.justmap.map.data;
+package ru.bulldog.justmap.map.data.classic;
 
 import ru.bulldog.justmap.client.JustMapClient;
 import ru.bulldog.justmap.client.config.ClientSettings;
+import ru.bulldog.justmap.map.data.IChunkData;
+import ru.bulldog.justmap.map.data.Layer;
 import ru.bulldog.justmap.network.ClientNetworkHandler;
 import ru.bulldog.justmap.util.DataUtil;
 import ru.bulldog.justmap.util.Dimension;
