+++ src/main/java/ru/bulldog/justmap/map/data/classic/WorldData.java
@@ -1,9 +1,12 @@
-package ru.bulldog.justmap.map.data;
+package ru.bulldog.justmap.map.data.classic;
 
 import ru.bulldog.justmap.client.config.ClientSettings;
 import ru.bulldog.justmap.event.ChunkUpdateEvent;
 import ru.bulldog.justmap.event.ChunkUpdateListener;
 import ru.bulldog.justmap.map.IMap;
+import ru.bulldog.justmap.map.data.IWorldData;
+import ru.bulldog.justmap.map.data.Layer;
+import ru.bulldog.justmap.map.data.RegionPos;
 import ru.bulldog.justmap.util.DataUtil;
 import ru.bulldog.justmap.util.math.MathUtil;
