+++ src/main/java/ru/bulldog/justmap/map/data/classic/WorldManager.java
@@ -1,4 +1,4 @@
-package ru.bulldog.justmap.map.data;
+package ru.bulldog.justmap.map.data.classic;
 
 import java.io.File;
 import java.util.ArrayList;
@@ -27,6 +27,7 @@ import ru.bulldog.justmap.enums.MultiworldDetection;
 import ru.bulldog.justmap.event.ChunkUpdateEvent;
 import ru.bulldog.justmap.event.ChunkUpdateListener;
 import ru.bulldog.justmap.map.IMap;
+import ru.bulldog.justmap.map.data.WorldKey;
 import ru.bulldog.justmap.util.DataUtil;
 import ru.bulldog.justmap.util.JsonFactory;
 import ru.bulldog.justmap.util.RuleUtil;
